// src/app/models/payroll/employee.model.ts
import { Resource } from '../shared/ressource.model';
import { Contract } from './contract.model';
import { EmployeeDocument } from './employee-document.model';
import { FileInfo } from '../shared/objectToSend';
import { Qualification } from './qualification.model';
import { PaymentType } from '../payment/payement-type.model';
import { EmployeeSkills } from './employee-skills.model';
import { JobEmployee } from './job-employee.model';
import { Review } from '../rh/Review.model';
import { Office } from '../shared/office.model';
import { Note } from './note.model';
import { EmployeeTrainingSession } from '../rh/employee-training-session.model';
import { Country } from '../administration/country.model';
import { City } from '../administration/city.model';
import { TimeSheet } from '../rh/timesheet.model';
import { Leave } from './leave.model';
import { ExitEmployee } from './exit-employee.model';
import { EmployeeState } from '../enumerators/employee-state.enum';

export class Employee extends Resource {
  Sex: number;
  FirstName: string;
  LastName: string;
  FullName: string;
  Email: string;
  Cin: string;
  Matricule: string;
  AddressLine1: string;
  AddressLine2: string;
  AddressLine3: string;
  AddressLine4: string;
  AddressLine5: string;
  PersonalPhone: string;
  ProfessionalPhone: string;
  Picture: string;
  BirthDate?: Date;
  BirthPlace: string;
  PseudoSkype: string;
  Facebook: string;
  Linkedin: string;
  SocialSecurityNumber: string;
  IdCitizenship: number;
  IdGrade: number;
  IsActive: boolean;
  Echelon: number;
  Category: string;
  HiringDate: Date;
  ResignationDate: Date;
  ResignationDepositDate: Date;
  IdCountry: number;
  IdCity: number;
  ZipCode: number;
  Contract: Array<Contract>;
  ChildrenNumber: number;
  FamilyLeader: boolean;
  CinAttached: string;
  DependentParent: number;
  IsForeign: boolean;
  ChildrenDisabled: number;
  ChildrenNoScholar: number;
  HomeLoan: number;
  IdPaymentType: number;
  Note: Array<Note>;
  EmployeeDocument: Array<EmployeeDocument>;
  PictureFileInfo: FileInfo;
  CinFileInfo: Array<FileInfo>;
  Rib: string;
  IdBank: number;
  IdUpperHierarchy: number;
  IdOffice: number;
  Qualification: Array<Qualification>;
  IdPaymentTypeNavigation: PaymentType;
  EmployeeSkills: Array<EmployeeSkills>;
  JobEmployee: Array<JobEmployee>;
  pictureEmployesSrc: any;

  ReviewIdEmployeeCollaboratorNavigation: Review[];
  ReviewIdManagerNavigation: Review[];
  IdOfficeNavigation: Office;
  EmployeeTrainingSession: Array<EmployeeTrainingSession>;
  IdCountryNavigation: Country;
  IdCitizenshipNavigation: Country;

  IdCityNavigation: City;
  Status: EmployeeState;
  HierarchyLevel: string;
  MaritalStatus?: number;
  EmployeeExit: Array<ExitEmployee>;
  TimeSheetIdEmployeeNavigation: TimeSheet[];
  LeaveIdEmployeeNavigation: Leave [];
  ResignedFromExitEmployee: boolean;
  image: any;
  CurrentContract: Contract;
}